Implemented enroling instructor page
The page can be accessed via a button located on the class list page
beside the import users button. On the page, the current instructors are
listed. To add new instructors, the user (instructor/admin) will find the
instructor in the dropdown menu and press Enrol.

Currently, only users that have an Instructor system role and are not
already in the course are listed in the dropdown menu. In the future,
I plan to have the enroling TAs and individual students on this page
as well. The api has already been setup to enrol other user types. Currently,
the role defaults to Instructors if no roles are given. For students and TAs,
we may want to implement any method to search for the user(s)

The UI is currrently very basic. Some things that I have to be changed include:
- icon for the enrol button on the classlist page
- button for the unenrol link on the enrol page
- some instructions/information on the enrol page

Courses are still showing up even after the user has been dropped from the course;
therefore I have modified the authorization class to ignore courses that have
type_dropped. I have also added this extra filter for the course list api.

- changed address for api to generate instructor/TA labels for answers and comments
- lastly, I've created tests for the new API I wrote