Fixed Issue #86
Previously, we just counted number of judgements made by a user
for each question for the number of evaluations completed. However,
this would not work when multiple criteria are used. To solve
this problem, I basically find a criteria of the question and return
number of judgements for that question. We assume the judgements
for all of the criteria for each answer pair is done at the same time.