Refactored delete for questions, answers, and comments
I have basically removed all the blank delete controllers and moved
the delete functionality to the controllers they are used, namely
CourseQuestionsController and QuestionViewController.

Also the pages are no longer reloaded to reflect the removal of an
object. Instead, the object is spliced from the array.

I have also modified the delete confirmation directive in the question
module to take the delete function and a keyword (eg. answer). The
keyword is optional. If no keyword is given, the message still makes sense.

For example:
- If "question" is passed, the message will be: "Are you sure you want to remove
this question?"
- If no keyword is passed, the message will be: "Are you sure you want to remove
this?"