implemented more meaningful display name generator
Before, if a display name is not provided we give the user a random
string as their display name. To make it more meaningful, I have
created a display name generator that appends the user's system role,
first name (if given), and a 4 digit number.

I have also removed the creation of dictionary from the import
algorithm, because it is not needed.