added new Criteria attribute
I added the default attribute to the model to let us know whether
the user wants to use it again in another course or just for the
current course. This idea came from a discussion about the possibility
of a user being overwhelmed with criteria, if every criteria they
have ever created is returned.