Refactored Evaluations page
- UI has been slightly refactored, more work is needed
- due to the refactoring, the answer author api function has been
removed (including its tests)
- removed answerpair api functions. Currently it is an empty model.
I've left it to create other api functions required for reports for
example.
- created an api function to return all of the students in the class
and its tests
- removed duplicate dataformat for CriteriaAndCourses
- refactored "notScoredEnd" filter to not use Javascript's filter
function
- created a function to query for the evaluations' answer pairs
when called from a link. This is so that we do not load all of the
data initially or else it may increase the loading time for
unnecessary data