Refactored class import
I've modified the api to only return error messages and number of
successful enrolment instead of returning all the details we can
generate - including the students that were dropped. We assume,
instructors may be overwhelmed with all the information and do not
need to know all these details.

So now, the user is redirected to the Manage Users view if there
are no errors. They are only redirected to the "Results" page if
there are errors