implemented enroling and unenroling users from groups
The group column in class list view now displays a column of dropdown
menus with the group the user is enroled in selected. If the user is not
in a group, the default value "-- None --" is selected.

To change the group, all the instructor needs to do is change the value
in the dropdown value. A success message will appear when it's done

To just remove them from a group, "-- None --" should be selected.
Again, a success message will appear.