refactored group member model
After discussion, it was decided that it would be better if the group
member model is related to Users instead of CoursesAndUsers. Since
the feature is still under development, I just dropped the
GroupsAndCoursesAndUsers model and created the GroupsAndUsers model.
Another reason was it was not possible to write a generic script to
drop the foreign key constraint for coursesandusers_id because we need
the name, which is autogenerated and may be different depending on the
SQL server being used

Other than minor refactoring to change the model used from
GroupsAndCoursesAndUsers to GroupsAndUsers and coursesandusers_id to
users_id, I also did the following:

- removed the requirement that the user has to be enroled in the course
to be dropped from a group. The reason is it also makes sense to drop
a user from a group even if they are no longer enroled in the course.
- moved the groups relationship to users, which will list all of the
active GroupsAndUsers object the user is related to.
- under CoursesAndUsers, I created a hybrid property that filters the
list of groups in Users to only list the groups in the course
- refactored tests to reflect the changes made