Reduce overall height of Questions screen for easier scrolling
Includes cutting off longer answers at 200px (with read more button to reveal full answer) and toggling replies to answers (with initial hidden display of all replies). 

One potential issue is if the user loads the screen in a large window, then shrinks that window. It seems unlikely a user would do this and not be able to re-expand the window, but this action may cut off longer answers w/o adding a read more button. (If this becomes an issue, future fixes could be watching the height of the div or simply using overflow:scroll and no read more button.)