Issue #111
I've initialized all the time periods. All of the time are set to
11:59pm.

Answer Start: today @ 11:59pm
Answer End: 7 days from today @ 11:59pm
Judge Start: same as Answer End
Judge End: 7 days from Answer End @ 11:59pm

I've also set the timepicker to display the minutes in steps of 1 instead
of 5. The reason is if we initialize the time to 11:59, then the steps
would be 49, 54, 59, etc.