Implement inline PDF viewing.
Inline PDF viewing is now implemented in the acj-pdf-inline directive.
All the instances where view pdf in a new window was used should've been
moved to this new directive.

The directive takes in a list of PDFs and a label string. Each entry in
the list of PDFs is expected to be a hash containing the name of the pdf
and the alias of the pdf. This directive can be imported as
'ubc.ctlt.acj.common.pdf'.

The directive uses viewerjs's iframe include to view PDFs. The directive
uses an isolate scope to make sure that multiple directives on a page
doesn't interfere with each other.

The old view pdf in a new window module has been removed. The attachment
module was not removed entirely as it seems they're used for uploading
too. It might be a good idea to move the attachment module to common,
since it's now a smaller module.

To prevent CSS conflicts in the question page, the modifications to the
list-inline class was changed to question-metadata-list. Since the PDF
directive also uses the list-inline class for listing PDFs, it was
changing the PDF file listing too.