Move status messages closer to screen center.
This widens the message to 90% width and 25% down from the top of the
screen.

Fixes #116

Toastr by default only has top and bottom messages. In order to have one
closer to the center of the screen, we need to add a custom one. This
customization is located in toast.less.