Add gulp task to compile and minify js libs, css.
For production, there is now a 'prod' task. It compiles and minifies
javascript libraries and the ACJ less files. It'll replace the includes
in index.html to the compiled and minified version.

Tried to do the same for the Angular modules, but it didn't work. The
order of the module includes didn't work out as some modules depended on
each other. There are workarounds for this, but they don't seem quite as
nice, so leaving it alone for now.