Ticket #121
Implemented student filter for the answers on the question view page

- modified the api for retrieving the students in a course to return
the students' fullname if they have permissions for coursesandusers and
display names if they do not have the permissions (eg. students).
	- reason: for the student filter (accessible by students)
	- modified test case to reflect this change
- modified the group members api to return "name" instead so that
it matches the structure for the classlist api
- implemented the student filter (author) which is similar to the
filter system on the evaluation results page
	- if the course has groups and have permissions to view them,
	a group filter will appear. It will not only filter the list
	of answers but also filter the list of students
	- visible for all users
	- added text that will display if no answers match the filters
- The "Answers" header has also been updated to show "Answers for Name",
if a name is selected in the filter.