Added Self-Evaluation column to Activity Screen
Currently, it only appears if self-evaluation option is selected.
Also it only display whether the student has completed the requirement
or not (assuming 1) with Yes or No.

Note: Currently assuming self-evaluation type to be no comparison

Also fixed gradebook not loading sometimes. The reason is the list
of students was not retrieved in time when we click on the Gradebook
tab. It seems to be fixed when I query for the list again in the
Gradebook controller.