Ticket #112 - Google Analytics
I have implemented the foundation for adding Google Analytics to
the application. The main Google Analytics javascript code has been
included in index.html. It would not do anything until the configuration
file is included as well, which I named tracking.js

All administrators have to do is follow the steps in the README file
to create the tracking.js file and run gulp `(nodemodules/.bin/gulp)
tracking` to include the file in index.html.

Note:
- if the application has already been installed (before this commit),
you will have to run `gulp` to install the dependency gulp-inject
- gulp-inject will not add 'tracking.js' to index.html if the file does
not exist.
- when committing index.html don't commit the line to include
tracking.js. I'm in the process of finding a way to ignore the one line
if it exists. The line is found in the Google Analytics section.