added 'round' column to PostsForAnswers to keep track of the
number of times the answer has been evaluated.

NOTE: database upgrade required
Currently one count is for each judgement, so if an evaluation has
two criteria, the answer will get two counts for each evaluation.
The main role of the new column is to attempt to ensure the
answes are evaluated a similar number of times. The ideal case
would of course be the same number of times for each answer.