refactored answer pairing algorithm
This is to keep the number of comparisons per answer evenly
distributed.
The answers are first grouped by the number of times they are judged.
Then we try to find the closest scored paris within the group
with the lowest number of comparisons. If a valid pair is not found,
we try to find a valid pair by merging the the next lowest number
of comparison group with the lowest number of comparison groups.