Ticket #176
Instead of using ng-if to remove and insert the evaluation period
form fields into the question form every time the checkbox
is checked or unchecked, I've changed it to use ng-show so that
the date formatting would not disappear. Since we are using ng-show,
I've also changed the fields to use ng-required to set whether those
datetime fields are required.