Fix #185, no feedback for non-existing CAS user
When user logs in from CAS, even though the user logged in correctly to
CAS server, the user may not exists in this app. An access denied
message shows up.

To achieve this in SPA, there are some challenges to make both backend
and frontend happy. The solutions is:
* After user logged in successfully in CAS, the user will be redirect
  back to a endpoint on the app, where handles ticket validation and
  user validation.
* If user exists, we login user. If not, we put a message in the session
  and redirect user to the homepage, where an /api/session call will be
  generated.
* Then frontend will generate a /api/session call. login_manager will
  intercept the call and pass it to unauthorized_handler, where an 403
  will be generated along with the message stored in session.
* Front end receives 403 and pop the message to the user

Because multiple session calls will be generated concurrently from the
homepage, it will result multiple 403 responses and duplicated pop-up
messages. The $http cache has been enabled to solve this issue.