Fixed HTTP 403 error's user feedback
When a user receives a 403 error, not only do they get a Toaster
error message, but the login dialog also appears. This gives an
incorrect impression to the user, because they were not logged
out of the system.

To fix this, I have changed the CAS error message to be formatted
the same way that Flask-Bouncer does. Additionally, I added a 'type'
attribute with the value 'CAS'.

Now, whenever a 403 is thrown, we will check whether it has the type
attribute and if it does check if the value is 'CAS'. If both conditions
are met, then the login dialog will appear.

Currently, only the first condition is necessary, because Flask-Bouncer's
response object does not have the attribute 'type'. However in the future
we may use the type attribute for other 'types' of 403 errors.