Fix inline pdf viewer not working in IE9/10.
Looks like ViewerJS is not compatible with IE9/10. Since pdf.js works
fine in IE9/10, I've just switched to using only pure pdf.js.

It turns out that not much needed to be changed to get pdf.js working.
Unfortunately, I had to modify pdf.js itself. There is a one line change
in web/viewer.js where the DEFAULT_URL is changed to take in the file we
want to show. Note that Mozilla does ask that we modify the viewer as a
curtesy, so I've made some cosmetic changes as well. This does mean that
we can't just drop in a fresh version of pdf.js for updates though.

Fixes #204