Fix 2+ editor instances on page break math preview
If there are multiple editor instances on a page, the math editor's
preview does not work. This is due to the preview div being retrieved by
static ID. When there are multiple editor instances, there are also
multiple preview divs. Which means that only the first preview div is
retrieved and modified, while the rest just remain blank.

Fix is to use CKEditor's automatic ID naming for the preview div. This
results in having to split the preview div's label into its own element
since the automatic ID is only given to the first element.