First commit for #198
- implemented $http caching for getting course (with id)
- created an interceptor to remove the cache if the object has been
modified or deleted.
- since the course objects are now cached, we no longer need a specific
API function to get the courses' names to improve efficiency. I have now
removed the function and its test cases.
- added cache for question (with id)
- Added caching for answer module
- answers' cache is cleared when an answer in the question is edited
or deleted
- enhanced the interceptor to remove cache for url without the trailing
id (if one exists)
- modified question view's answer tab to use $scope.answers.length to
get current count of answers. The reason is after an answer is deleted
the answers_count variable that the question model has is not updated.
- refactored enrol/unenrol api url to be simpler and to be the same
as the other urls in the ClassList resource
- caching view api functions to simplify some of the code
- refactored CourseResource.getQuestions to just use QuestionResource.get
so that the invalidation of the cached question would work more smoothly