Fixed Issue #243
We are now using a seed which will force the same pair to be
returned to the student if they refresh the page if the pool of
answers do not change. If the pool changes since the last the student
grabbed a pair, the new pair they get may change. The seed is generated
by multiplying their user id with the round they are on, so that they
will use a different seed at each round