Refactor and optimize to reduce SQLs
This commit set expire_on_commit to False. It solves the issue that
objects in session expiring after activity event fired (an INSERT SQL
and commit), which causes SQLAlchemy refresh the objects from DB when
doing the marshal. It reduced the amount of SQLs almost half. However,
it may cause the objects in session outdated. So be careful to access
the objects that may outdated after committing.