Modified Score Algorithm
Moved scoring algorithm to submodule. Module allows additional algorithms to be added in the future (more scoring algorithms, pair selection algorithm).

Also the calculate score function should be called for each criteria set of comparisons since the scores for each criteria are independent of each other.

Due to separating out and recoding parts of the algorithm two issues where corrected.

1) The number of rounds recorded in the scores table is now correct (was incrementing round count by 1 x number of criteria for question)

2) The excepted score of an answer now correctly takes into account comparisons where the answer never wins against the opponent. This causes the overall excepted score to be higher than before for any answer that lost all comparisons against an opponent.

example: A vs B with A having 0 wins and 1 lose
exp(0 - 1)/(1+ exp(0 - 1)) ~= 0.27
so the expected score for A will be ~0.27 instead of 0 after one round with A losing