Cleanup API
- Removed system and course role API endpoints (front-end no longer needs them)
- Consolidated comparison count/available and answer count/answered API calls into two new calls in the assignment API (/api/courses/course_id/assignments/status and /api/courses/course_id/assignments/assignment_id/status)

The comparison and answer counts & boolean endpoints are all currently used in the context of the current user and for asking questions on the user's status for an assignment (have I answered the assignment?, can I make comparisons for the assignment?, have I completed by comparisons for the assignment?). The idea is to put these related calls into a more general purpose endpoint to help answer these questions and reduce the amount of nested API calls in the front-end. It should also be more easily maintainable and extendable.

I intend to extend the new endpoints to include information of answer drafts in an assignment.

Related #258