Convert tab indention to 4 spaces
All python files were using spaces instead of tabs already so instead of converting all of them to tabs, I went the opposite way and converted the .js and .html files into using spaces as well. Note that the .js and .html files also had a mix of tabs and spaces for indention. The dev.md has also been updated to reflect this change