Edit default criterion for new assignments
Public criterion are still added to the criteria list in a new assignment. If the user selects to edit it, they will generate a new criterion instead. When saving a new assignment, duplicates are automatically generated for public criterion still in the list. This allows users to later freely edit the text of criterion based on public criterion

Existing assignments using default criterion are still not be editable them

Fixes: #431