Add global unique identifiers
Simplifies LTI and CAS/SAML logins by automatically creating accounts on login (when global unique identifier is set for that login method). When account is created, global unique identifier will be associated with the user so that when logging in with a different method, account can be automatically linked.

If the LTI consumer does not use global unique identifiers, then it can be linked to a compair account though login (via username/password or CAS/SAML) as per normal.

- Simplified how xapi external actor accounts are handled. Now when properly setup will use global unique identifier (no need to check `third_party_user` or `lti_user` tables)
- Changed `lti_consumer`'s`user_id_override` to `global_unique_identifier_param` and removed all `lti_user` records linked to consumer that used it. Users will automatically be relinked to these accounts on thier next lti launch or membership refresh since the old lti user_id field is automatically stored as the user's global unique identifier. this change was done so that user id would be properly stored for using the LTI Assignment and Grades Service later on.
- Added user management function `generate_global_unique_identifiers` to automatically populate global unique identifiers based on existing CAS, SAML, and/or LTI links (requires having environemnt variables already setup for CAS and SAML and should not be needed for LTI due to migration but might be helpful anyways).
- Added Index on `comparable` in `answer` table
- Also fix acceptance tests

Closes #606