Improve Kaltura media ownership options
When using global_unique_identifiers there is now the option to set the Kaltura media owner_id to the user's global_unique_identifiers instead of the system administrators. This shifts ComPAIR from asking like a bucket account to allowing users to manage their own media. This could cause some potential issues if the user deleted or changes accessibility to the media.

Related to #713