Update AWS readme to fix issue with DB creation
The current line (44) to retrieve the app container ID is incorrect at it contains the wrong name for the instance.  In addition, just correcting the name retrieves two running containers as they have the same instance name.  I've updated the readme so that it pulls the single running app container instance.