ADD pylti library locally
Bring in the pylti library locally so that we can modify it to support
some Canvas extension parameters. This commit is the stock pylti
library. I've removed the django contrib as we're not using it and it
was causing runtime issues, we're not using django anyways.

Also brought in pylti required libraries into requirements.txt.

Note that this is an LTI 1.1 library, and as such is being deprecated in
the future. Since it's a dead end library, we're not too worried about
upstream changes anymore, hence why this hacky fix.