EDIT LTIOutcome sends submittedAt timestamp
Canvas uses the time it receives grades as the time that the student
submitted the grade. This means that if there's a delay on the ComPAIR
side in sending the grade to Canvas, Canvas thinks the student's submit
time is late. This happened when the job queue worker encountered some
issues and got stuck. Now we have some student grades wrongly marked as
late.

To avoid this, Canvas supports an extension parameter for sending the
submit time. This is adding an element under the <replaceResultRequest>
element in the form of:
<submissionDetails>
    <submittedAt>2017-04-16T18:54:36.736+00:00</submittedAt>
</submissionDetails>

The timestamp is an ISO 8601 timestamp that requires timezone
information. Unfortunately, we do not store the created/modified time
with timezone information. So we have to make some assumptions here. We
assume that the ComPAIR server has the same timezone as MariaDB. So we
apply the ComPAIR server's timezone to the timestamp.

Note that we are taking the modified time as the submit time. It turns
out that the regular grade sync is done as a course grade. And this
course grade gets modified every time a student does an assignment. So
the course_grade actually stores the most *recent* grade for an
assignment for a student, not a course_grade as the name implies. This
means that the created timestamp would be wildly inaccurate, since a
course grade seems to be created when a student first enter the system.