ADD btn to download all attachments in assignment
This button is in the assignment's "Participation" tab table under the
"Attachments" table header.

The group and student filter in the "Participation" tab does not apply
to the download all attachments button. So the button is hidden when
filters are in effect.

The attachments are all gathered into a single zip file. The zip is
named <assignment name> [attachments-<assignment UUID>].zip. I've
allowed some level of compression to the zip file, in order to reduce
file sizes. This is zlib's default compression level of 6, which is
apparently a good compromise between time taken & file size reduction.
Hopefully the time is not too egregious. A test in an assignment with 76
attachments showed a relatively quick few seconds spent compressing.
While waiting for this to complete, the button shows a spinner and is
disabled.

The actual attachment files in the zip is named after the uploader in
the format: <name> - <student number> - <uuid file name>
Student number is skipped if user doesn't have one. The uuid file name
is the name ComPAIR gave to the actual file in the storage system.

The zip file is saved in the persistent reports directory. I don't think
there's enough access control protection in place for reports, but since
it's been in place for years and is out of scope for this feature, it
will have to be dealt with later.

If assignment does not have attachments, a message stating as such is
returned.