ADD test case for download all attachments api
Test case to make sure that the attachments zip file is properly
created.

One thing to note is that the api call limits answers only to those that
are active and from users who are still enroled in the course. This was
a bit clunky to take care of in the test case.

Added the Pillow library for creating images. This is used to create
real files for the test case.