FIX celery upgrade to v5.2.3
Both kombu and redis also needs to be updated due to new requirements
brought on by the new celery version. I've set them to the current
latest. It doesn't seem to break anything luckily.

The biggest breaking celery change is to the cli option order. Global
options, such as --app, needs to be placed after the celery command and
cannot be placed after sub commands (like worker). So I've fixed the
celery invocations I've found. Note that --loglevel needs to be after
the sub command, it's unrecognized as a global option apparently.