FIX tests broken by serialization change in celery
The serialization change is basically that python tuples, when put into
the celery queue, no longer gets converted back as a python tuple. It
gets converted to a list.

Not sure where to put the blame for this, but Celery uses Kombu for
serialization, and that uses msgpack. I'm guessing it's some change
in those dependencies. A quick look through their change logs didn't see
anything particulary concerning, but it could be side effect of those
libraries ditching support for python 2.