FIX task set_passwords broken by celery serialize
The values are all there and correct, why is it not working?!

Turns out this is part of the celery upgrade's more strict adherence to
the JSON spec. In JSON, object keys are all strings, they cannot be
ints. So, when we serialized the user id, using is a key, it went in as
a int, and came out as a string. I guess previously, it went in as an
int and came out as an int. When we tried to get the new user password,
with the user id as an int, "None" came back cause there's no int keys.

Just needed to convert the user id to a string to match the new key
type.