UPGRADE sqlalchemy from 1.3.8 to 1.4.32
Fix warnings about scalar_subquery() has to be explicitly declared.
E.g.: a lot of the virtual fields that establish counts of something
does a select query that returns a scalar value, so needs to be
explicitly declared as a subquery.

Passing a subquery into the in_() construct resulted in a warning that
the subquery gets coerced into a select(). Turns out the warning is
basically just saying that you don't have to pass in a subquery and can
just leave it as a select() in a roundabout way. So I've deleted the
subquery() conversion in these cases to leave them as plain select().

join() using the table name as a string throws an error now, e.g.: if
you try to join("answer"), it tells you 'Entity namespace for "mapped
class Answer->answer" has no property "answer"'. Passing in the actual
model class Answer fixes it.

Also updated sqlalchemy associated packages like pymysql & alembic.