FIX cannot import name 'json' from 'itsdangerous'
The json api in itsdangerous was deprecated and then removed in the
latest version. So need to pin it to an older version.

While researching this, found out that soft_unicode from markupsafe also
has the same issue, so I've pinned that as well.

Both can be removed if we ever upgrade to flask 2.