EDIT github action deploy fail on http errors
Github action calls on our gitlab pipeline for actual deployment. It
uses curl to call the api trigger, but curl doesn't exit with an error
when it gets an http error code. You need to add a --fail to make curl
exists with an error. The downside is that we lose the body of the error
response, so additional error messages from the server are lost. But at
least this will alert us to an error in the pipeline.