EDIT replace sqlalchemy-enum34 w sqlalchemy enum
enum34 is a legacy artifact, as python enums were introduced in python
3.4, but ComPAIR used to support earlier python versions. As these
earlier python versions have been end-of-life for years now, we can now
drop the enum34 usage and just use the standard enum support.

While enum34 doesn't seem to be used directly anymore, it looks like
sqlalchemy-enum34 is still used for all enum columns in the models.

Unfortunately, sqlalchemy-enum34.EnumType behaves differently from
sqlalchemy.Enum, so it's not a drop-in replacement.

* sqlalchemy-enum34's EnumType stores the enum's value to the database.
* sqlalchemy's Enum stores the enum's key to the database.

So we have an alembic migration that'll convert between these two
behaviours. The migration is a bit complicated, as each column has to be
converted in multiple steps to the end result due to the db collation
being case insensitive and to sqlalchemy's constraint enforcement. But
it should be reversible, so if issues crop up, it should be
downgradeable.

Note that since the models now use sqlalchemy.Enum, they will error if
the migration is not run.