UPGRADE cleanly upgradeable python dependencies
Most notably, upgraded python 3.7 to 3.8. 3.7 still has a year of
support, but moving up now since it might take us a while to get this
out anyways.

Used pip-review to find out packages that needed to be upgraded. This
commit only covers those dependencies that can be cleanly upgraded with
no issues.

Flask needs to be upgraded to flask 2, but that's pretty serious
upgrade, so it's not part of this commit.

Flask-Login breaks with newer versions.

Flask-Bouncer also changed the behaviour of methods so that they throw
exceptions by default. We need to change the permission checking method
to use the new ones that don't throw exceptions where we don't expect
it.

Various packages were kept back due to conflicts with newer libs.

Also updated the README for more detail on running the test cases.