UPGRADE frontend test to node 16, run on chrome
The tests were running on PhantomJS, which hasn't been maintained for
years apparently. I tried to run the tests locally and it was erroring
out, though it was erroring with a different error than the error in
github actions. In either case, it would be a good idea to run out
PhantomJS as the cause, so I'm switching it to try to run on Chrome.

The 'puppeteer' package provides a headless Chromium which
can replace it. So I've switched the dependencies to run on headless
Chrome.

I've also updated the github action for the frontend test to the latest
setup-node and set it to run on the more current node 16.