FIX frontend test can't find tincan lib
The earlier frontend test fix resulted in a different error. Now it
couldn't find the tincan library. I'm guessing because Github Actions
doesn't like that we're installing the tincan library from a local repo
directory. So I've forked the tincan library and put it in a public repo
so it can download like all the other bower libs.