UPGRADE celery config to celery 4+ lower case ones
Support for the all upper case config values will be dropped in Celery
6, so this is future proofing too.

Note that all config passed to Compair are still all upper case. It's
just that we'll convert it to lower case when passing it to Celery.
These env vars are first read by Compair which then pass it to Celery.

To add a Celery configuration values, convert the lower case Celery
setting into uppercase and prepend it with a `CELERY_` prefix, and then
add it to the right overrideable dict in `configuration.py`. E.g.:
`broker_url` becomes `CELERY_BROKER_URL`. The setting will be stripped
of the `CELERY_` prefix and converted back to lower case via Flask
Config's `get_namespace()` before being fed to Celery.

Another example, CELERYBEAT_SCHEDULE is Celery 3's config, in Celery 4+,
this becomes beat_schedule. So to pass it into Compair, you use
CELERY_BEAT_SCHEDULE.

This commit is a prerequisite before doing #983