ADD celery worker max task and max mem options
We've noticed memory leaks on long running worker tasks. Which
eventually causes the worker to die due to out-of-memory. To prevent
this, added support for Celery's worker_max_tasks_per_child and
worker_max_memory_per_child options.

They need to be set in Compair as env vars:

* CELERY_WORKER_MAX_TASKS_PER_CHILD - Takes an int. Kills a worker
process and forks a new one when it has executed the given number of
tasks

* CELERY_WORKER_MAX_MEMORY_PER_CHILD - Takes an int as memory in
kilobytes. Kills a worker process and forks a new one when it hits the
given memory usage, the currently executing task will be allowed to
complete before being killed.

They seem to work in testing. It looks like the main worker process
forks a pool of task workers (default is # of cores on system). It's
these forked task workers that gets killed, not the main worker process.
So when I generate tasks that reach the given set limit, I can see a
database connection gets broken and the 'ForkPoolWorker-<int>' count
gets incremented on subsequent tasks.

This fixes #983