FIX lti launch to assignment not working
There's two causes for this:
- assignment needs to be allowed as a parameter
- Canvas doesn't include the url query part when generating signatures

Looks like one of the updates we pulled for the lti library added
parameter verification and we need to add assignment to the allowed
list.

Assignment uid is specified as a query parameter after the launch url.
The lti library takes the entire launch url, including this assignment
parameter, when validating the signature. However, it looks like Canvas
only uses the original launch url (without the assignment uid query) to
generate the signature. So as a quick fix, we'll pass in only the url
without queries to the signature validator. This fix wouldn't work if
our original launch url contains other queries, but afaik, ComPAIR
shouldn't encounter this issue.